files: separate discretization from refinement
